# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Taskts",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is Task app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://timemanagement-back.onrender.com"
    }
  ],
...
  "paths": {
    "/tasks": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get all Tasks",
        "operationId": "getTasksController",
        "description": "Get list of users with this endpoint",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "example": 1
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "example": 10
            }
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string",
              "example": "sortOrder"
            },
            "description": "All Task's fields. Default value - _id"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ],
              "example": "asc"
            }
          },
          {
            "in": "query",
            "name": "taskType",
            "schema": {
              "type": "string",
              "enum": [
                "work",
                "home",
                "personal"
              ],
              "example": "work"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found Tasks!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found Tasks!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/tasks"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 2
                        },
                        "perPage": {
                          "type": "number",
                          "example": 4
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 4
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 4
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": false
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "post": {
        "tags": [
          "Tasks"
        ],
        "summary": "Create task",
        "operationId": "createTask",
        "description": "Create a task with payload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "task",
                  "timeDeclaration",
                  "taskType",
                  "date",
                  "status"
                ],
                "properties": {
                  "task": {
                    "definitions": "task name",
                    "type": "string",
                    "example": "dddhh aayyz"
                  },
                  "timeDeclaration": {
                    "definitions": "task's time, min",
                    "type": "number",
                    "example": 120
                  },
                  "date": {
                    "definitions": "task's date & time",
                    "type": "date",
                    "example": "2025-09-10T18:10:00.000+00:10"
                  },
                  "isFavourite": {
                    "definitions": "Would you like or not?",
                    "type": "boolean",
                    "example": false
                  },
                  "taskType": {
                    "definitions": "task's work place",
                    "type": "string",
                    "example": "work",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "default": "personal"
                  },
                  "status": {
                    "definitions": "task's status",
                    "type": "string",
                    "example": "pending",
                    "enum": [
                      "completed",
                      "pending",
                      "working"
                    ],
                    "default": "personal"
                  },
                  "timeReal": {
                    "definitions": "task's time, min",
                    "type": "number",
                    "example": 120
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a task!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/tasks"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get stask by id",
        "operationId": "getTaskIDController",
        "description": "Get Task by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6692528c222b4be7b4b17ebd"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get Task by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found Task with id 6692528c222b4be7b4b17ebd!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/tasks"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "taskRouterTaskstags": [
          "Tasks"
        ],
        "summary": "Delete task",
        "operationId": "deleteTask",
        "description": "Delete a task from DB",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6692528c222b4be7b4b17ebd"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Delete task by id"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "patch": {
        "tags": [
          "Tasks"
        ],
        "summary": "Change task by ID",
        "operationId": "ChangeTask",
        "description": "Change a task by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "6692528c222b4be7b4b17ebd"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "task": {
                    "definitions": "task name",
                    "type": "string",
                    "example": "dddhh aayyz"
                  },
                  "timeDeclaration": {
                    "definitions": "task's time, min",
                    "type": "number",
                    "example": 120
                  },
                  "date": {
                    "definitions": "task's date & time",
                    "type": "date",
                    "example": "2025-09-10T18:10:00.000+00:10"
                  },
                  "isFavourite": {
                    "definitions": "Would you like or not?",
                    "type": "boolean",
                    "example": false
                  },
                  "taskType": {
                    "definitions": "task's work place",
                    "type": "string",
                    "example": "work",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "default": "personal"
                  },
                  "status": {
                    "definitions": "task's status",
                    "type": "string",
                    "example": "pending",
                    "enum": [
                      "completed",
                      "pending",
                      "working"
                    ],
                    "default": "personal"
                  },
                  "timeReal": {
                    "definitions": "task's time, min",
                    "type": "number",
                    "example": 120
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Change a task",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully changed a task!"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/tasks"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "tasks": {
        "type": "object",
        "required": [
          "_id",
          "task",
          "timeDeclaration",
          "timeReal",
          "taskType",
          "date",
          "status",
          "userId"
        ],
        "properties": {
          "_id": {
            "definitions": "task's ID",
            "type": "string",
            "example": "6692528c222b4be7b4b17ebd"
          },
          "task": {
            "definitions": "task",
            "type": "string",
            "example": "dddhh aayyz"
          },
          "timeDeclaration": {
            "definitions": "120",
            "type": "number",
            "example": 120
          },
          "timeReal": {
            "definitions": "120",
            "type": "number",
            "example": 120
          },
          "isFavourite": {
            "definitions": "Would you like or not?",
            "type": "boolean",
            "example": false
          },
          "taskType": {
            "definitions": "task's work place",
            "type": "string",
            "example": "work",
            "enum": [
              "work",
              "home",
              "personal"
            ],
            "default": "personal"
          },
          "userId": {
            "definitions": "Autorth's task ID",
            "type": "string",
            "example": "6692528c222b4be7b4b17ebd"
          },
          "date": {
            "definitions": "date",
            "type": "date",
            "example": "2025-09-13T18:10:00.000Z"
          },
          "status": {
            "definitions": "task's status",
            "type": "string",
            "example": "pending",
            "enum": [
              "completed",
              "pending",
              "working"
            ],
            "default": "pending"
          }
        }
      },
...
    },
    "responses": {
...
      "404": {
        "description": "Unauthorized error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Route not found"
                },
                "data": {
                  "type": "object",
                  "required": [
                    "message"
                  ],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Route not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}